fix PaymentList and Payment migrate and model and fix middleware routs

+++ routes/api.php
@@ -12,6 +12,7 @@ use \App\Http\Controllers\UserController;
 use \App\Http\Controllers\PaymentController;
 use \App\Http\Controllers\CardController;
 use \App\Http\Controllers\ReportController;
+
 /*
 |--------------------------------------------------------------------------
 | API Routes
@@ -25,58 +26,57 @@ use \App\Http\Controllers\ReportController;
 
 Route::group(['middleware' => ['auth:sanctum', 'CheckUserStatus']], function () {
 
-//Dashboard
-
-    Route::get('/Admin/Dashboard', [ReportController::class, 'Dashboard'])->name('Dashboard');
 
 //Auth
-    Route::post('/Register/GetInformation', [RegisterController::class, 'GetInformation'])->name('GetInformation');
+    Route::post('/Register/GetInformation', [RegisterController::class, 'GetInformation'])->name('GetInformation')->middleware(['permission:GetInformation']);
     Route::post('/Logout', [RegisterController::class, 'Logout'])->name('Logout');
+    Route::post('/Register/Admin/UpdatePassword', [RegisterController::class, 'UpdatePassword'])->name('UpdatePassword')->middleware(['permission:Admin.UpdatePassword']);
+
+
+//Dashboard
+    Route::get('/Admin/Dashboard', [ReportController::class, 'Dashboard'])->name('Dashboard')->middleware(['permission:Dashboard']);
 
 
 // Customer
-    Route::put('/Customer/Edit/{id}', [UserController::class, 'Edit'])->name('EditCustomer');
-    Route::put('/Customer/Update', [UserController::class, 'Update'])->name('UpdateCustomer');
-    Route::delete('/Customer/Delete/{id}', [UserController::class, 'Delete'])->name('DeleteCustomer');
-    Route::get('/Customer/List', [UserController::class, 'List'])->name('ListCustomer');
+    Route::put('/Customer/Edit/{id}', [UserController::class, 'Edit'])->name('EditCustomer')->middleware(['permission:Customer.Edit']);
+    Route::put('/Customer/Update', [UserController::class, 'Update'])->name('UpdateCustomer')->middleware(['permission:Customer.Update']);
+    Route::delete('/Customer/Delete/{id}', [UserController::class, 'Delete'])->name('DeleteCustomer')->middleware(['permission:Customer.Delete']);
+    Route::get('/Customer/List', [UserController::class, 'List'])->name('ListCustomer')->middleware(['permission:User.BlockOrActive']);
     Route::get('/User/operation/{id}', [UserController::class, 'operation'])->name('operation')->middleware(['permission:User.operation', 'CheckOrderAccess']);
     Route::get('/User/BlockOrActive/{id}', [UserController::class, 'BlockOrActive'])->name('BlockOrActive')->middleware(['permission:User.BlockOrActive']);
 
 
 //Product
-    Route::post('/Product/Create', [ProductController::class, 'Create'])->name('CreateProduct');
-    Route::put('/Product/Edit/{id}', [ProductController::class, 'Edit'])->name('EditProduct');
-    Route::put('/Product/show/Edit', [ProductController::class, 'showEdit'])->name('EditProduct');
-    Route::get('/Product/List', [ProductController::class, 'List'])->name('ListProduct');
-    Route::delete('/Product/Delete/{id}', [ProductController::class, 'Delete'])->name('DeleteProduct');
-
+    Route::post('/Product/Create', [ProductController::class, 'Create'])->name('CreateProduct')->middleware(['permission:Product.Create']);
+    Route::put('/Product/Edit/{id}', [ProductController::class, 'Edit'])->name('EditProduct')->middleware(['permission:Product.Edit']);
+    Route::put('/Product/show/Edit', [ProductController::class, 'showEdit'])->name('showEditProduct')->middleware(['permission:Product.showEdit']);
+    Route::get('/Product/List', [ProductController::class, 'List'])->name('ListProduct')->middleware(['permission:Product.List']);
+    Route::delete('/Product/Delete/{id}', [ProductController::class, 'Delete'])->name('DeleteProduct')->middleware(['permission:Product.Delete']);
 
 //Order
-    Route::post('/Order/Create', [OrderController::class, 'Create'])->name('CreateOrder');
+    Route::post('/Order/Create', [OrderController::class, 'Create'])->name('CreateOrder')->middleware(['permission:Order.Create']);
     Route::put('/Order/Edit/{id}', [OrderController::class, 'Edit'])->name('EditOrder')->middleware(['permission:Order.Edit', 'checkOrderPermission']);
-    Route::put('/Order/Show/Edit', [OrderController::class, 'ShowEdit'])->name('ShowEditOrder');
-    Route::get('/Order/List', [OrderController::class, 'List'])->name('ListOrder');
-    Route::delete('/Order/Delete/{id}', [OrderController::class, 'Delete'])->name('DeleteOrder');
+    Route::put('/Order/Show/Edit', [OrderController::class, 'ShowEdit'])->name('ShowEditOrder')->middleware(['permission:Order.ShowEdit']);
+    Route::get('/Order/List', [OrderController::class, 'List'])->name('ListOrder')->middleware(['permission:Order.List']);
+    Route::delete('/Order/Delete/{id}', [OrderController::class, 'Delete'])->name('DeleteOrder')->middleware(['permission:Order.Delete']);
 
 
 //Media
-    Route::post('/UploadImage/{id}', [ProductController::class, 'UploadImage'])->name('UploadImage');
-    Route::post('/UploadMainImage/{id}', [ProductController::class, 'UploadMainImage'])->name('UploadMainImage');
-    Route::post('/UploadVideo/{id}', [ProductController::class, 'UploadVideo'])->name('UploadVideo');
+    Route::post('/Order/UploadImage/{id}', [ProductController::class, 'UploadImage'])->name('UploadImage')->middleware(['permission:Order.UploadImage']);
+    Route::post('/Order/UploadMainImage/{id}', [ProductController::class, 'UploadMainImage'])->name('UploadMainImage')->middleware(['permission:Order.UploadMainImage']);
+    Route::post('/Order/UploadVideo/{id}', [ProductController::class, 'UploadVideo'])->name('UploadVideo')->middleware(['permission:Order.UploadVideo']);
 
 
 //Payment
     Route::post('/Payment', [PaymentController::class, 'Payment'])->name('Payment');
+    Route::post('Admin/PaymentList', [PaymentController::class, 'PaymentList'])->name('PaymentList')->middleware(['permission:PaymentList']);
+    Route::post('Admin/PaymentFilter', [PaymentController::class, 'PaymentFilter'])->name('PaymentFilter')->middleware(['permission:PaymentFilter']);
 
 // Card
-    Route::get('Card/showUserTickets', [CardController::class, 'showUserTickets'])->name('showUserTickets');
-    Route::get('Card/showAllTickets', [CardController::class, 'showAllTickets'])->name('showAllTickets');
-    Route::get('Card/download-pdf/{id}', [CardController::class, 'downloadPdf'])->name('download_pdf');
-    Route::get('Card/FilterCard', [CardController::class, 'FilterCard'])->name('FilterCard');
-
-
-
-
+    Route::get('Card/UserTickets', [CardController::class, 'UserTickets'])->name('UserTickets')->middleware(['permission:Card.UserTickets']);
+    Route::get('Card/AllTickets', [CardController::class, 'AllTickets'])->name('AllTickets')->middleware(['permission:Card.AllTickets']);
+    Route::get('Card/DownloadPdf/{id}', [CardController::class, 'DownloadPdf'])->name('DownloadPdf')->middleware(['permission:Card.DownloadPdf']);
+    Route::get('Card/FilterCard', [CardController::class, 'FilterCard'])->name('FilterCard')->middleware(['permission:Card.FilterCard']);
 
 
     //Article
@@ -101,4 +101,3 @@ Route::post('/Register/GetCodeSent', [RegisterController::class, 'GetCodeSent'])
 Route::post('/Register/Admin/Login', [RegisterController::class, 'AdminLogin'])->name('AdminLogin');
 Route::post('/Register/Admin/EmailPassword', [RegisterController::class, 'EmailPassword'])->name('EmailPassword');
 Route::post('/Register/Admin/VerifyCode', [RegisterController::class, 'VerifyCode'])->name('VerifyCode');
-Route::post('/Register/Admin/UpdatePassword', [RegisterController::class, 'UpdatePassword'])->name('UpdatePassword');
